changes after mail

+++ src/main/java/com/project/spring/repositories/FilmRepository.java
@@ -4,6 +4,6 @@ import com.project.spring.entities.Film;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FilmRepository extends CrudRepository<Film, Long> {
+public interface FilmRepository extends CrudRepository<Film, Integer> {
 
 }
